+++ console/src/features/RelationshipsTable/components/TableCell.tsx
@@ -3,8 +3,11 @@ import { FaTable, FaColumns } from 'react-icons/fa';
 
 const ColumnCell = ({ columnName }: { columnName: ReactText }) => (
   <>
-    <FaColumns className="fill-current text-sm text-muted mr-1" />
-    {columnName}
+    <FaColumns
+      className="fill-current text-sm text-muted mr-1"
+      title="Column"
+    />
+    <span className="mr-2">{columnName}</span>
   </>
 );
 
@@ -17,7 +20,7 @@ const TableCell = ({
 }) => (
   <>
     <div className="flex items-center">
-      <FaTable className="fill-current text-sm text-muted mr-1" />
+      <FaTable className="fill-current text-sm text-muted mr-1" title="Table" />
       {tableName}
       <span className="px-2">/</span>
       {cols ? cols.map(i => <ColumnCell columnName={i} />) : null}